Added cache headers for .html files
"This code snippet checks if a page has been modified since it was last displayed. If so, it sends a "304 not modified" header and exits, otherwise the content is rendered." source: http://css-tricks.com/snippets/php/intelligent-php-cache-control/

Adds Last-Modified, Etag (using md5; performance is still good for files under 1MB), and Cache-Control header.